pyproject.toml
[project]
name = "LinguaScreen"
version = "0.1.0"
description = "AI Apps"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-translation-text>=1.0.1",
    "azure-ai-vision-imageanalysis>=1.0.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "openai>=1.82.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
